cookbooks/python/agent-builder/output-vibecoder/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
evolve-sdk>=0.1.0

